web/requirements.txt
Flask==0.12.2
requests==2.18.4
jsonschema==2.6.0
gunicorn==19.7.1
psycopg2-binary==2.7.4
pandas==0.22.0
